# Backend/mediaplan/Swagger.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8083","description":"Generated server url"}],"paths":{"/api/pacing-alerts/{id}/status":{"put":{"tags":["pacing-alert-controller"],"operationId":"updateStatus","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/StatusUpdateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponsePacingAlertResponse"}}}}}}},"/api/media-plans/{id}":{"get":{"tags":["media-plan-controller"],"operationId":"getById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseMediaPlanResponse"}}}}}},"put":{"tags":["media-plan-controller"],"operationId":"update","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MediaPlanRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseMediaPlanResponse"}}}}}},"delete":{"tags":["media-plan-controller"],"operationId":"delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}}}}}},"/api/media-plans/{id}/status":{"put":{"tags":["media-plan-controller"],"operationId":"updateStatus_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/StatusUpdateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseMediaPlanResponse"}}}}}}},"/api/line-items/{id}":{"get":{"tags":["media-line-item-controller"],"operationId":"getById_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseMediaLineItemResponse"}}}}}},"put":{"tags":["media-line-item-controller"],"operationId":"update_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MediaLineItemRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseMediaLineItemResponse"}}}}}},"delete":{"tags":["media-line-item-controller"],"operationId":"delete_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}}}}}},"/api/line-items/{id}/status":{"put":{"tags":["media-line-item-controller"],"operationId":"updateStatus_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/StatusUpdateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseMediaLineItemResponse"}}}}}}},"/api/insertion-orders/{id}/status":{"put":{"tags":["insertion-order-controller"],"operationId":"updateStatus_3","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/StatusUpdateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseInsertionOrderResponse"}}}}}}},"/api/pacing-alerts/run":{"post":{"tags":["pacing-alert-controller"],"operationId":"runNow","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseMapStringInteger"}}}}}}},"/api/media-plans":{"get":{"tags":["media-plan-controller"],"operationId":"getAll","parameters":[{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponsePagedResponseMediaPlanResponse"}}}}}},"post":{"tags":["media-plan-controller"],"operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MediaPlanRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseMediaPlanResponse"}}}}}}},"/api/media-plans/{planId}/line-items":{"get":{"tags":["media-line-item-controller"],"operationId":"getByPlan","parameters":[{"name":"planId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListMediaLineItemResponse"}}}}}},"post":{"tags":["media-line-item-controller"],"operationId":"create_1","parameters":[{"name":"planId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MediaLineItemRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseMediaLineItemResponse"}}}}}}},"/api/insertion-orders":{"get":{"tags":["insertion-order-controller"],"operationId":"getAll_1","parameters":[{"name":"publisherId","in":"query","required":false,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListInsertionOrderResponse"}}}}}},"post":{"tags":["insertion-order-controller"],"operationId":"create_2","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InsertionOrderRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseInsertionOrderResponse"}}}}}}},"/api/delivery-records":{"post":{"tags":["delivery-record-controller"],"operationId":"record","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeliveryRecordRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseDeliveryRecordResponse"}}}}}}},"/api/pacing-alerts":{"get":{"tags":["pacing-alert-controller"],"operationId":"getAlerts","parameters":[{"name":"status","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListPacingAlertResponse"}}}}}}},"/api/line-items/{lineItemId}/delivery":{"get":{"tags":["delivery-record-controller"],"operationId":"getByLineItem","parameters":[{"name":"lineItemId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListDeliveryRecordResponse"}}}}}}},"/api/insertion-orders/{id}":{"get":{"tags":["insertion-order-controller"],"operationId":"getById_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseInsertionOrderResponse"}}}}}}}},"components":{"schemas":{"StatusUpdateRequest":{"type":"object","properties":{"status":{"type":"string","minLength":1}},"required":["status"]},"ApiResponsePacingAlertResponse":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/PacingAlertResponse"},"timestamp":{"type":"string","format":"date-time"}}},"PacingAlertResponse":{"type":"object","properties":{"alertId":{"type":"integer","format":"int32"},"lineItemId":{"type":"integer","format":"int32"},"alertType":{"type":"string"},"alertDate":{"type":"string","format":"date"},"pacingPercent":{"type":"number"},"status":{"type":"string"},"createdAt":{"type":"string","format":"date-time"}}},"MediaPlanRequest":{"type":"object","properties":{"briefId":{"type":"integer","format":"int32"},"plannerId":{"type":"integer","format":"int32"},"totalBudgetAllocated":{"type":"number","minimum":0.0},"channelMix":{"type":"string"},"startDate":{"type":"string","format":"date"},"endDate":{"type":"string","format":"date"}},"required":["briefId","endDate","plannerId","startDate","totalBudgetAllocated"]},"ApiResponseMediaPlanResponse":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/MediaPlanResponse"},"timestamp":{"type":"string","format":"date-time"}}},"MediaPlanResponse":{"type":"object","properties":{"planId":{"type":"integer","format":"int32"},"briefId":{"type":"integer","format":"int32"},"plannerId":{"type":"integer","format":"int32"},"totalBudgetAllocated":{"type":"number"},"channelMix":{"type":"string"},"startDate":{"type":"string","format":"date"},"endDate":{"type":"string","format":"date"},"status":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},"MediaLineItemRequest":{"type":"object","properties":{"channel":{"type":"string","minLength":1},"publisher":{"type":"string","minLength":1},"format":{"type":"string"},"plannedImpressions":{"type":"integer","format":"int32","minimum":1},"plannedBudget":{"type":"number","minimum":0.0},"cpm":{"type":"number"},"flightStart":{"type":"string","format":"date"},"flightEnd":{"type":"string","format":"date"}},"required":["channel","flightEnd","flightStart","plannedBudget","plannedImpressions","publisher"]},"ApiResponseMediaLineItemResponse":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/MediaLineItemResponse"},"timestamp":{"type":"string","format":"date-time"}}},"MediaLineItemResponse":{"type":"object","properties":{"lineItemId":{"type":"integer","format":"int32"},"planId":{"type":"integer","format":"int32"},"channel":{"type":"string"},"publisher":{"type":"string"},"format":{"type":"string"},"plannedImpressions":{"type":"integer","format":"int32"},"plannedBudget":{"type":"number"},"cpm":{"type":"number"},"flightStart":{"type":"string","format":"date"},"flightEnd":{"type":"string","format":"date"},"status":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},"ApiResponseInsertionOrderResponse":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/InsertionOrderResponse"},"timestamp":{"type":"string","format":"date-time"}}},"InsertionOrderResponse":{"type":"object","properties":{"ioId":{"type":"integer","format":"int32"},"lineItemId":{"type":"integer","format":"int32"},"publisherId":{"type":"integer","format":"int32"},"orderDate":{"type":"string","format":"date"},"startDate":{"type":"string","format":"date"},"endDate":{"type":"string","format":"date"},"committedImpressions":{"type":"integer","format":"int32"},"orderValue":{"type":"number"},"status":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},"ApiResponseMapStringInteger":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{"type":"object","additionalProperties":{"type":"integer","format":"int32"}},"timestamp":{"type":"string","format":"date-time"}}},"InsertionOrderRequest":{"type":"object","properties":{"lineItemId":{"type":"integer","format":"int32"},"publisherId":{"type":"integer","format":"int32"},"orderDate":{"type":"string","format":"date"},"startDate":{"type":"string","format":"date"},"endDate":{"type":"string","format":"date"},"committedImpressions":{"type":"integer","format":"int32","minimum":1},"orderValue":{"type":"number","minimum":0.0}},"required":["committedImpressions","endDate","lineItemId","orderDate","orderValue","publisherId","startDate"]},"DeliveryRecordRequest":{"type":"object","properties":{"lineItemId":{"type":"integer","format":"int32"},"reportingDate":{"type":"string","format":"date"},"deliveredImpressions":{"type":"integer","format":"int32","minimum":0},"spend":{"type":"number","minimum":0.0}},"required":["deliveredImpressions","lineItemId","reportingDate","spend"]},"ApiResponseDeliveryRecordResponse":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/DeliveryRecordResponse"},"timestamp":{"type":"string","format":"date-time"}}},"DeliveryRecordResponse":{"type":"object","properties":{"deliveryId":{"type":"integer","format":"int32"},"lineItemId":{"type":"integer","format":"int32"},"reportingDate":{"type":"string","format":"date"},"deliveredImpressions":{"type":"integer","format":"int32"},"spend":{"type":"number"},"createdAt":{"type":"string","format":"date-time"}}},"ApiResponseListPacingAlertResponse":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{"type":"array","items":{"$ref":"#/components/schemas/PacingAlertResponse"}},"timestamp":{"type":"string","format":"date-time"}}},"Pageable":{"type":"object","properties":{"page":{"type":"integer","format":"int32","minimum":0},"size":{"type":"integer","format":"int32","minimum":1},"sort":{"type":"array","items":{"type":"string"}}}},"ApiResponsePagedResponseMediaPlanResponse":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/PagedResponseMediaPlanResponse"},"timestamp":{"type":"string","format":"date-time"}}},"PagedResponseMediaPlanResponse":{"type":"object","properties":{"content":{"type":"array","items":{"$ref":"#/components/schemas/MediaPlanResponse"}},"page":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"last":{"type":"boolean"}}},"ApiResponseListMediaLineItemResponse":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{"type":"array","items":{"$ref":"#/components/schemas/MediaLineItemResponse"}},"timestamp":{"type":"string","format":"date-time"}}},"ApiResponseListDeliveryRecordResponse":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{"type":"array","items":{"$ref":"#/components/schemas/DeliveryRecordResponse"}},"timestamp":{"type":"string","format":"date-time"}}},"ApiResponseListInsertionOrderResponse":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{"type":"array","items":{"$ref":"#/components/schemas/InsertionOrderResponse"}},"timestamp":{"type":"string","format":"date-time"}}},"ApiResponseVoid":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{},"timestamp":{"type":"string","format":"date-time"}}}}}}